% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apportionment.R
\name{.build_divisors}
\alias{.build_divisors}
\title{Build the divisor vector for a given method}
\usage{
.build_divisors(method, seats)
}
\description{
Build the divisor vector for a given method
}
\keyword{internal}
