% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{.seatsIdeal}
\alias{.seatsIdeal}
\title{Compute Ideal Seat Shares}
\usage{
.seatsIdeal(ne, nd, np, sample)
}
\arguments{
\item{ne}{Integer: number of elections.}

\item{nd}{Integer: number of districts.}

\item{np}{Integer: number of parties.}

\item{sample}{A list generated by \code{sampleElectionData}.}
}
\value{
A \code{data.frame} with columns: \code{elec}, \code{SeatTotalIdeal},
  \code{SeatShareIdeal}, \code{Party}.
}
\description{
Computes the proportional ("ideal") seat shares at country level for each
party in each election.
}
\keyword{internal}
