% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{if.parties.null}
\alias{if.parties.null}
\title{Generate Party Identifiers}
\usage{
if.parties.null(x)
}
\arguments{
\item{x}{Integer: the number of parties.}
}
\value{
A character vector of length \code{x} with deterministic 3-letter codes.
}
\description{
Create 3-letter identifiers for parties when no party labels are supplied.
}
\examples{
if.parties.null(4)

}
