% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{Disp2}
\alias{Disp2}
\title{Aggregate-Level Disproportionality Measures}
\usage{
Disp2(seed = 1000, np = 3, nd = 1, ne = 100, dist = "lnorm",
  minTS = 3, lim = 16, threshold = 0, threshold_country = 0, ...)
}
\description{
The function computes aggregate-level measures of disproportionality. It also models and plots relationships between values of aggregate-level disproportionality measures and district sizes.
}
