% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{if.parties.null}
\alias{if.parties.null}
\title{If Parties NULL}
\usage{
if.parties.null(x)
}
\value{
3-letter identifiers for parties.
}
\description{
Create 3-letter identifiers for parties if 'parties' parameter in 'divisorMethods' or 'LR.Hamilton' functions is NULL.
}
\keyword{internal}
