% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apportionment.R
\name{LR_Hamilton}
\alias{LR_Hamilton}
\title{Hamilton-Hare Largest Remainder Method}
\usage{
LR_Hamilton(parties = NULL, votes = NULL, seats = NULL, threshold = 0,
  order_name = TRUE, ...)
}
\arguments{
\item{parties}{A character vector for party labels in the same order as in the numerical vector of votes for parties. If NULL, 3-letter identifiers (e.g. 'XYZ') will be assigned.}

\item{votes}{A numeric vector of votes received by each party.}

\item{seats}{The number of seats to be apportioned.}

\item{threshold}{A numeric value between [0, 1]. Default is set to 0.}

\item{order_name}{A logical value. Sort rows of a dataframe with results by party name (alphabetically).}
}
\value{
A data.frame of length parties containing apportioned seats.
}
\description{
Allocate seats to parties using Hamilton-Hare largest remainder apportionment method.
}
\examples{
(...)
}
