% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{plot_Disp}
\alias{plot_Disp}
\title{Plot Disproportionality Measures Related to District Sizes and Numbers of Parties}
\usage{
plot_Disp(bias_data, tse = c(0, 5/12, -1/12, -4/12), ...)
}
\description{
The function plots the relationships betweenon, on the one hand, values of a variaty of disproportionality measures and, on the other hand, district sizes or numbers of parties.
}
