% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{plot_Disp2}
\alias{plot_Disp2}
\title{Plot Aggregate-Level Disproportionality Measures}
\usage{
plot_Disp2(data = NULL, methods = c("DH", "SL", "H", "Imperiali"))
}
\description{
The function plots relationships between values of aggregate-level disproportionality measures and district sizes.
}
