% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{sampleElectionData}
\alias{sampleElectionData}
\title{Sample Election Data}
\usage{
sampleElectionData(seed = 0, dist = "uniform", np, nd, ne, mean = NULL,
  sd = NULL, rate = NULL, max, TS, formula_dist, ...)
}
\arguments{
\item{seed}{A scalar value (a random seed) used to initialize a pseudorandom number generator.}

\item{dist}{A character value for a probability distribution function used to sample data. 'Uniform' (default) = 'uniform', 'Exponential' = 'exp' and 'Log-Normal' = 'lnorm' distributions are available.}

\item{np}{A numeric value for the number of parties.}

\item{nd}{A numeric value for the number of electoral districts.}

\item{ne}{A numeric value for the number of elections.}

\item{mean}{A numeric value for the location parameter (a log-normal distribution).}

\item{sd}{A numeric value for the scale parameter (a log-normal distribution).}

\item{rate}{A numeric value for the rate parameter (an exponential distribution).}

\item{max}{A numeric value for the upper limit of the domain for a truncated log-normal distribution used to sample votes for parties at the constituency level.}

\item{TS}{A numeric value for the total number of seats apportioned among electoral district (if nd (the number of district) > 1), using a selected electoral formula.}

\item{formula_dist}{A character value that specifies the apportionment method used to divide seats ('TS') among electoral districts.}
}
\value{
A list containing:
Votes_Dist_Party,
Seats_Dist,
Votes_Share_Party,
Votes_Total_Dist,
Votes_Total_Party,
Votes_Total.
}
\description{
Return a pseudorandom sample (from a truncated log-normal distribution) of n election results using a variaty of paremeters such as: the number of parties (np), the total number of seats (TS) or the number of electoral districts.
}
\examples{
(...)
}
