% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{simulate_Disp}
\alias{simulate_Disp}
\title{Per-party Disproportionality Measures for Varying District Sizes and Numbers of Parties}
\usage{
simulate_Disp(seed = 0, dist = "lnorm", np, nd = 1, ne, rate = 1/25000,
  mean = 10, sd = 1.2, max = 1e+05, formula, formula_dist = "hh",
  threshold = 0, threshold_country = 0, minTS = 3, maxTS = 20,
  jump = 2, ...)
}
\description{
The function computes per-party disproportionality measures using simulated election data.
}
