% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{simulate_E}
\alias{simulate_E}
\title{Simulate Elections under Proportional Representation}
\usage{
simulate_E(seed, dist = "lnorm", np, nd, ne, mean, sd, rate, max, TS, formula,
  formula_dist, threshold, threshold_country)
}
\value{
A list containg the following data objects:
seat_excess,
apportionment,
disp,
summary.
}
\description{
The function simulate election results and compute a variaty of disproportionality measures.
}
\examples{
(...)
}
